Fix review comment: moved cron job, added logger, fixed entity with lombok annotations, moved changelog, and created preCondition in changelog for Liquibase to not create new table if one exists
