+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -4,6 +4,5 @@ import com.example.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    // Можно добавлять кастомные методы, например:
     User findByName(String name);
 }
