+++ src/main/resources/application.properties
@@ -11,3 +11,4 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.yaml
+app.cleanup-cron=*/10 * * * * *
