+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,19 +1,3 @@
 databaseChangeLog:
-  - changeSet:
-      id: 1
-      author: demo
-      changes:
-        - createTable:
-            tableName: users
-            columns:
-              - column:
-                  name: id
-                  type: BIGSERIAL
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: name
-                  type: VARCHAR(255)
-                  constraints:
-                    nullable: false
+  - include:
+      file: db/changelog/001-create-users.yaml
